增加评论功能测试脚本，新增聊天功能接口

+++ app/web/comments.py
@@ -14,29 +14,27 @@ class AddComment(Resource):
         提交评论 和 回复评论
         params: commentSeqid  评论seqid
         params: commentText
-        params: isPublic
         params: articlesId
         '''
-        commentSeqid = request.args.get('commentSeqid')
-        commentText = request.args.get('text')
-        isPublic = request.args.get('isPublic')
-        relationArtId = request.args.get('articlesId')
+        commentSeqid = request.form.get('commentSeqid')
+        commentText = request.form.get('text')
+        relationArtId = g.artSeqid
         
         # 回复评论
         if commentSeqid:
-            _tmpRes = g.tableComment.add_comment(commentText, isPublic, relationArtId, commentSeqid)
+            _tmpRes = g.tableComment.add_comment(g.user.seqid, commentText, g.isPublic, relationArtId, commentSeqid)
 
         else:
-            _tmpRes = g.tableComment.add_comment(commentText, isPublic, relationArtId)
+            _tmpRes = g.tableComment.add_comment(g.user.seqid, commentText, g.isPublic, relationArtId)
 
         if _tmpRes:
             g.retMsg['status'] = 1
             g.retMsg['code'] = 200
+            
+            # 更新动态表评论数
             if not g.tableArticle.update_article(relationArtId, comments = 1):
-                
                 g.retMsg['msg'] = '动态表评论数修改失败'  
         else:
-            g.retMsg['code'] = 403
             g.retMsg['msg'] = '评论提交失败'
 
         return jsonify(g.retMsg)
@@ -50,30 +48,34 @@ class DeleteComment(Resource):
         params: commentSeqid  评论seqid
         params: articlesId
         '''
-        seqid = request.args.get('commentSeqid')
-        relationArtId = request.args.get('articlesId')
+        seqid = int(request.args.get('commentSeqid'))
+        relationArtId = g.artSeqid
+        commentDict = g.tableComment.get_comment(seqid)
+        
+        if g.user.seqid != commentDict.get('userid'):
+            g.retMsg['msg'] = '只能删除自己的评论'
+            return jsonify(g.retMsg)
+
         _tmpRes = g.tableComment.delete_comment(seqid)
         if _tmpRes:
             g.retMsg['status'] = 1
             g.retMsg['code'] = 200
             if not g.tableArticle.update_article(relationArtId, comments = -1):
-                g.retMsg['code'] = 403
                 g.retMsg['msg'] = '动态表评论数修改失败'  
-
         else:
-            g.retMsg['code'] = 403
             g.retMsg['msg'] = '评论删除失败'
         return jsonify(g.retMsg)
 
 class GetComment(Resource):
     '''
     获取评论
-    params: commentSeqid  评论seqid
-    params: articlesId  动态seqid
+    *params: commentSeqid  评论seqid
+    *params: articlesId  动态seqid
     '''
+    @check_token
     def get(self):
         commentSeqid = request.args.get('commentSeqid')
-        relationArtId = request.args.get('articlesId')
+        relationArtId = g.artSeqid
         if commentSeqid:
             _tmpDict = g.tableComment.get_comment(commentSeqid)
             if _tmpDict:
@@ -81,7 +83,6 @@ class GetComment(Resource):
                 g.retMsg['code'] = 200
                 g.retMsg['data'] = _tmpDict
             else:
-                g.retMsg['code'] = 403
                 g.retMsg['msg'] = '评论获取失败'
 
         elif relationArtId:
@@ -91,7 +92,6 @@ class GetComment(Resource):
                 g.retMsg['code'] = 200
                 g.retMsg['data'] = _tmpDict
             else:
-                g.retMsg['code'] = 403
                 g.retMsg['msg'] = '评论获取失败'
         
         return jsonify(g.retMsg)
